backend/requirements.txt
# FastAPI and web framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.0
asyncpg>=0.25.0

# Redis
redis==5.0.1
aioredis==2.0.1

# HTTP client
httpx==0.25.2
requests==2.31.0

# AI and LangChain
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.7
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Time series and financial calculations
yfinance==0.2.28
ccxt==4.1.77

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-watch==4.2.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Security
cryptography>=39.0.0,<42.0.0
bcrypt>=4.0.0

# CORS
fastapi-cors==0.0.6
